chore(RbacRulesService): move RbacServiceData
