+++ src/app/[locale]/settings/_components/role-settings/FormMappingHelper.ts
@@ -1,4 +1,4 @@
-import { BaSyxRbacRule, TargetInformation, rbacRuleTargets } from 'lib/types/RbacServiceData';
+import { BaSyxRbacRule, TargetInformation, rbacRuleTargets } from 'lib/services/rbac-service/types/RbacServiceData';
 import {
     ArrayOfIds,
     RuleFormModel,
