+++ src/app/[locale]/settings/_components/role-settings/target-information/TargetInformationForm.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { Box, FormControl, MenuItem, Select, Typography } from '@mui/material';
-import { rbacRuleTargets } from 'lib/types/RbacServiceData';
+import { rbacRuleTargets } from 'lib/services/rbac-service/types/RbacServiceData';
 import { useTranslations } from 'next-intl';
 import { WildcardOrStringArrayInput } from 'app/[locale]/settings/_components/role-settings/target-information/WildcardOrStringArrayInput';
 import { Control, Controller, UseFormGetValues, UseFormSetValue } from 'react-hook-form';
